% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{get_change_points}
\alias{get_change_points}
\title{Get vector for sorting observations into terminal nodes}
\usage{
get_change_points(df, x)
}
\arguments{
\item{df}{dataframe}

\item{x}{covariates}
}
\value{
a vector
}
\description{
Get vector for sorting observations into terminal nodes
}
