% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{log_marginal_likelihood}
\alias{log_marginal_likelihood}
\title{Computes the log marginal likelihood for accepting/rejecting BART trees}
\usage{
log_marginal_likelihood(node_partial_res, kappa, omega)
}
\arguments{
\item{node_partial_res}{matrix}

\item{kappa}{scalar}

\item{omega}{matrix}
}
\description{
Computes the log marginal likelihood for accepting/rejecting BART trees
}
