% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{matrnorm}
\alias{matrnorm}
\title{Sampling from a matrix normal distribution MV(mu, U, V)}
\usage{
matrnorm(mu, U, V)
}
\arguments{
\item{mu}{mean matrix (real nxp matrix)}

\item{U}{scale matrix (positive-definite real nxn matrix)}

\item{V}{scale matrix (positive-definite real pxp matrix)}
}
\value{
one sample from a MV(mu, U, V) distribution
}
\description{
Sampling from a matrix normal distribution MV(mu, U, V)
}
\examples{
matrnorm(matrix(0, nrow=5, ncol=2), diag(5), diag(2))
}
