% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{MH}
\alias{MH}
\title{Metropolis-Hastings sampler}
\usage{
MH(p1, l1, p2, l2, q1 = 0, q2 = 0)
}
\arguments{
\item{p1}{scalar or vector}

\item{l1}{scalar or vector}

\item{p2}{scalar or vector}

\item{l2}{scalar or vector}

\item{q1}{scalar or vector}

\item{q2}{scalar or vector}
}
\value{
scalar or vector for accepting/rejecting proposed values
}
\description{
Metropolis-Hastings sampler
}
