% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missBART2.R
\name{missBART}
\alias{missBART}
\title{Title}
\usage{
missBART(
  x,
  y,
  x_predict = NA,
  n_trees = 20,
  burn = 100,
  iters = 100,
  thin = 2,
  predict = TRUE,
  tree_prior_params = tree_list(),
  hypers = hypers_list(),
  scale = TRUE,
  include_x = TRUE,
  include_y = TRUE,
  show_progress = TRUE,
  progress_every = 10,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
