% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missBART2.R
\name{missBART2}
\alias{missBART2}
\title{Title}
\usage{
missBART2(
  x,
  y,
  x_predict = NA,
  n_reg_trees = 20,
  n_class_trees = 20,
  burn = 100,
  iters = 100,
  thin = 2,
  predict = TRUE,
  tree_prior_params = tree_list(),
  hypers = hypers_list(),
  scale = TRUE,
  include_x = TRUE,
  include_y = TRUE,
  show_progress = TRUE,
  progress_every = 10,
  ...
)
}
\arguments{
\item{x}{covariates}

\item{y}{response}

\item{x_predict}{out-of-sample covariates. If not specificied, the default is set to NA and no out-of-sample predictions will be made.}

\item{n_reg_trees}{number of BART trees}

\item{n_class_trees}{number of probit BART trees}

\item{burn}{burn-in samples}

\item{iters}{post-burn-in samples}

\item{thin}{thinning}

\item{predict}{make out-of-sample predictions?}

\item{tree_prior_params}{prior parameters for BART trees}

\item{hypers}{prior parameters for BART parameters}

\item{scale}{scale data?}

\item{include_x}{Include x in probit model?}

\item{include_y}{Include y in probit model?}

\item{show_progress}{logical}

\item{progress_every}{integer value stating how often to update the progress bar.}

\item{...}{Catches unused arguments}
}
\description{
Title
}
