% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{multi_rMVN}
\alias{multi_rMVN}
\title{Sampling multiple times from a multivariate normal distribution}
\usage{
multi_rMVN(mean_mat, precision)
}
\arguments{
\item{mean_mat}{matrix containing n mean vectors}

\item{precision}{precision matrix}
}
\value{
a matrix where each row corresponds to a draw from a MVN distribution with mean equal to the rows of \code{mean_mat} and precision \code{precision}
}
\description{
Sampling multiple times from a multivariate normal distribution
}
\examples{
multi_rMVN(matrix(0, ncol=2, nrow=10), diag(2))
}
