% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBART.R
\name{mvBART}
\alias{mvBART}
\title{Title}
\usage{
mvBART(
  x,
  y,
  x_predict = NA,
  n_trees = 50,
  burn = 1000,
  iters = 1000,
  thin = 3,
  predict = TRUE,
  tree_prior_params = tree_list(),
  hypers = hypers_list(),
  scale = TRUE,
  show_progress = TRUE,
  progress_every = 10,
  ...
)
}
\arguments{
\item{x}{covariates}

\item{y}{response}

\item{x_predict}{out-of-sample covariates}

\item{n_trees}{number of trees}

\item{burn}{burn-in samples}

\item{iters}{post burn-in samples}

\item{thin}{thinning}

\item{predict}{whether or not to make out-of-sample predictions}

\item{tree_prior_params}{prior parameters for BART trees}

\item{hypers}{prior parameters for BART parameters}

\item{scale}{logical. Whether to scale data to range (-0.5, 0.5).}

\item{show_progress}{logical.}

\item{progress_every}{integer value stating how often to update the progress bar.}

\item{...}{Catches unused arguments}
}
\description{
Title
}
