% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{node_priors}
\alias{node_priors}
\title{Compute tree priors at the node level}
\usage{
node_priors(depth, prior_alpha, prior_beta)
}
\arguments{
\item{depth}{scalar}

\item{prior_alpha}{scalar}

\item{prior_beta}{scalar}
}
\description{
Compute tree priors at the node level
}
