% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{rMVN}
\alias{rMVN}
\title{Sampling once from a MVN(mu, Q^{-1}) distribution (Keefe)}
\usage{
rMVN(mu, Q)
}
\arguments{
\item{mu}{mean vector}

\item{Q}{precision matrix}
}
\value{
One sample from a MVN(mu, Q^{-1}) distribution
}
\description{
Sampling once from a MVN(mu, Q^{-1}) distribution (Keefe)
}
\examples{
rMVN(c(0,0), diag(2))
}
