% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{rMVNc}
\alias{rMVNc}
\title{Sampling from MVN(bQ^{-1},Q^{-1})}
\usage{
rMVNc(b, Q, is_chol = FALSE)
}
\arguments{
\item{b}{a vector}

\item{Q}{precision matrix}

\item{is_chol}{logical}
}
\value{
One sample from a MVN(bQ^{-1},Q^{-1}) distribution
}
\description{
Sampling from MVN(bQ^{-1},Q^{-1})
}
\examples{
rMVNc(c(0,0), diag(2))
}
