% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data_friedman}
\alias{sim_data_friedman}
\title{Simulate}
\usage{
sim_data_friedman(n, p = 1, scale_par = 1, scale_err = 0.5)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of responses}

\item{scale_par}{scale parameter}

\item{scale_err}{scale error}
}
\description{
Simulate
}
\examples{
data = sim_data_friedman(n = 1000, p = 3)
true_trees_data = data$true_trees
y = y_original = data$y
x = data$x
ome = data$ome
}
