% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{sim_kappa}
\alias{sim_kappa}
\title{MCMC sample of kappa}
\usage{
sim_kappa(mu, a, b)
}
\arguments{
\item{mu}{list of mu's from all BART trees in the current iteration}

\item{a}{hyperparameter}

\item{b}{hyperparameter}
}
\value{
one MCMC sample for kappa
}
\description{
MCMC sample of kappa
}
