% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_missing}
\alias{sim_missing}
\title{Simulating missing values from a probit regression model}
\usage{
sim_missing(
  x,
  y,
  include_x = FALSE,
  include_y = FALSE,
  min_missing_prop = 0.6,
  max_missing_prop = 0.9,
  ...
)
}
\arguments{
\item{x}{regression covariates}

\item{y}{regression response}

\item{include_x}{logical. Include x in probit model?}

\item{include_y}{logical. Include y in probit model?}

\item{min_missing_prop}{minimum proportion of missingness}

\item{max_missing_prop}{maximum proportion of missingness}

\item{...}{Catches unused arguments}
}
\description{
Simulating missing values from a probit regression model
}
