% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_missing_probit}
\alias{sim_missing_probit}
\title{Simulating missing values from a probit regression model}
\usage{
sim_missing_probit(
  x,
  y,
  include_x = FALSE,
  include_y = FALSE,
  min_missing_prop = 0.6,
  max_missing_prop = 0.9,
  ...
)
}
\arguments{
\item{x}{regression covariates}

\item{y}{regression response}

\item{include_x}{logical. Include x in probit model?}

\item{include_y}{logical. Include y in probit model?}

\item{min_missing_prop}{minimum proportion of missingness}

\item{max_missing_prop}{maximum proportion of missingness}

\item{...}{Catches unused arguments}
}
\value{
A list with the following components
\describe{
\item{B}{probit coefficients}
\item{m}{missingness indicators}
\item{missing_y}{inputted y with missing values deleted}
\item{missing_prop}{proportion of missingness per response variable}
\item{missing_index}{indices of missing cells in \code{missing_y}}
\item{obs_index}{indices of observed cells in \code{missing_y}}
\item{corR}{correlation matrix}
\item{zmod}{probit latent variable}
}
}
\description{
Simulating missing values from a probit regression model
}
\examples{
data <- sim_data_friedman(n = 1000, p = 3)
missing <- sim_missing_probit(x = x, y = y, include_x = TRUE, include_y = TRUE,
                              max_missing_prop = 0.99, min_node = 10,
                              min_missing_prop = 0.7, intercept = TRUE)
}
