% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_missing_trees}
\alias{sim_missing_trees}
\title{Simulate missing data from a BART model}
\usage{
sim_missing_trees(
  x,
  y,
  trees = 1,
  include_x = FALSE,
  include_y = TRUE,
  min_missing_prop = 0.6,
  max_missing_prop = 0.9,
  ...
)
}
\arguments{
\item{x}{regression covariates}

\item{y}{regression response}

\item{trees}{number of trees}

\item{include_x}{logical. Include x in probit model?}

\item{include_y}{logical. Include y in probit model?}

\item{min_missing_prop}{minimum proportion of missingness}

\item{max_missing_prop}{maximum proportion of missingness}

\item{...}{Catches unused arguments}
}
\value{
A list with the following components
\describe{
\item{m}{missingness indicators}
\item{missing_y}{inputted y with missing values deleted}
\item{missing_prop}{proportion of missingness per response variable}
\item{missing_index}{indices of missing cells in \code{missing_y}}
\item{obs_index}{indices of observed cells in \code{missing_y}}
\item{true_trees}{trees}
\item{z}{latent variable}
}
}
\description{
Simulate missing data from a BART model
}
\examples{
# data <- sim_data_trees(n = 100, p = 3, q = 4, trees = 6)
# missing <- sim_missing_trees(x = data$x, y = data$y, include_x = FALSE,
#                              include_y = TRUE, max_missing_prop = 0.99,
#                              min_node = 5, min_missing_prop = 0.7, trees = 1)
}
