% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{sim_mu}
\alias{sim_mu}
\title{MCMC sample of mu for all terminal nodes of a BART tree}
\usage{
sim_mu(change_points, partial_res, kappa, omega)
}
\arguments{
\item{change_points}{vector}

\item{partial_res}{matrix}

\item{kappa}{hyperparameter}

\item{omega}{precision matrix}
}
\value{
mu for all terminal nodes of a tree
}
\description{
MCMC sample of mu for all terminal nodes of a BART tree
}
