% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{sim_mu_node}
\alias{sim_mu_node}
\title{Sampling mu at tree terminal node level}
\usage{
sim_mu_node(node_partial_res, kappa, omega)
}
\arguments{
\item{node_partial_res}{matrix of partial residuals from trees}

\item{kappa}{hyperparameter}

\item{omega}{precision matrix}
}
\value{
a single mu (scalar or vector) at terminal node level
}
\description{
Sampling mu at tree terminal node level
}
