% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{sim_omega}
\alias{sim_omega}
\title{MCMC sample of BART precision matrix}
\usage{
sim_omega(y, y_hat, alpha, Vinv)
}
\arguments{
\item{y}{data}

\item{y_hat}{sum of BART outputs}

\item{alpha}{hyperparameter}

\item{Vinv}{hyperparameter}
}
\value{
one MCMC sample for BART precision matrix
}
\description{
MCMC sample of BART precision matrix
}
