% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{tree_priors}
\alias{tree_priors}
\title{Compute tree priors}
\usage{
tree_priors(nodes, parents, depth, prior_alpha, prior_beta)
}
\arguments{
\item{nodes}{nodes of trees}

\item{parents}{parents of nodes}

\item{depth}{current depth of tree}

\item{prior_alpha}{hyperparameter}

\item{prior_beta}{hyperparameter}
}
\description{
Compute tree priors
}
