% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{update_B}
\alias{update_B}
\title{MCMC sample of B, the matrix of probit parameters}
\usage{
update_B(
  x,
  y,
  z,
  tau_b = 0.01,
  include_x = TRUE,
  include_y = TRUE,
  center = TRUE
)
}
\arguments{
\item{x}{regression covaraites}

\item{y}{data}

\item{z}{latent variables}

\item{tau_b}{hyperparameter}

\item{include_x}{logical}

\item{include_y}{logical}

\item{center}{logical}
}
\value{
one MCMC sample of B
}
\description{
MCMC sample of B, the matrix of probit parameters
}
