% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{update_Psi}
\alias{update_Psi}
\title{MCMC sample of \eqn{\Psi} from \cite{Talhouk, A., Doucet, A., & Murphy, K. (2012). Efficient Bayesian inference for multivariate probit models with sparse inverse correlation matrices. Journal of Computational and Graphical Statistics, 21(3), 739-757.}}
\usage{
update_Psi(nu, S, B, R)
}
\arguments{
\item{nu}{degrees of freedom}

\item{S}{scale matrix}

\item{B}{matrix of probit parameters}

\item{R}{correlation matrix}
}
\value{
one MCMC sample of \eqn{\Psi}
}
\description{
MCMC sample of \eqn{\Psi} from \cite{Talhouk, A., Doucet, A., & Murphy, K. (2012). Efficient Bayesian inference for multivariate probit models with sparse inverse correlation matrices. Journal of Computational and Graphical Statistics, 21(3), 739-757.}
}
