% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{update_RB}
\alias{update_RB}
\title{MCMC sample of R (correlation matrix) and B (probit parameters)}
\usage{
update_RB(Sigma, gamma)
}
\arguments{
\item{Sigma}{matrix}

\item{gamma}{matrix}
}
\value{
one MCMC sample of R and B
}
\description{
MCMC sample of R (correlation matrix) and B (probit parameters)
}
