% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{update_sigma_gamma}
\alias{update_sigma_gamma}
\title{MCMC sample of \eqn{\Sigma} and \eqn{\gamma}, the latent variables in probit regression introduced by \cite{Talhouk, A., Doucet, A., & Murphy, K. (2012). Efficient Bayesian inference for multivariate probit models with sparse inverse correlation matrices. Journal of Computational and Graphical Statistics, 21(3), 739-757.}}
\usage{
update_sigma_gamma(p, Y, Psi, W)
}
\arguments{
\item{p}{scalar}

\item{Y}{matrix}

\item{Psi}{matrix}

\item{W}{matrix}
}
\value{
one MCMC sample of \eqn{\Sigma} and \eqn{\gamma}
}
\description{
MCMC sample of \eqn{\Sigma} and \eqn{\gamma}, the latent variables in probit regression introduced by \cite{Talhouk, A., Doucet, A., & Murphy, K. (2012). Efficient Bayesian inference for multivariate probit models with sparse inverse correlation matrices. Journal of Computational and Graphical Statistics, 21(3), 739-757.}
}
