% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{update_W}
\alias{update_W}
\title{MCMC sample of W, the latent variable in probit regression introduced by \cite{Talhouk, A., Doucet, A., & Murphy, K. (2012). Efficient Bayesian inference for multivariate probit models with sparse inverse correlation matrices. Journal of Computational and Graphical Statistics, 21(3), 739-757.}}
\usage{
update_W(R, z)
}
\arguments{
\item{R}{correlation matrix}

\item{z}{probit latent variables}
}
\value{
one MCMC sample of W
}
\description{
MCMC sample of W, the latent variable in probit regression introduced by \cite{Talhouk, A., Doucet, A., & Murphy, K. (2012). Efficient Bayesian inference for multivariate probit models with sparse inverse correlation matrices. Journal of Computational and Graphical Statistics, 21(3), 739-757.}
}
