% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{update_y_miss_reg}
\alias{update_y_miss_reg}
\title{MCMC sample of missing data}
\usage{
update_y_miss_reg(
  x,
  y_hat,
  m,
  y,
  z,
  B,
  R,
  omega,
  include_x = TRUE,
  include_y = TRUE
)
}
\arguments{
\item{x}{regression covariates}

\item{y_hat}{sum of BART outputs}

\item{m}{matrix of missing data indicator}

\item{y}{data}

\item{z}{probit regression latent variables}

\item{B}{matrix of probit parameters}

\item{R}{correlation matrix}

\item{omega}{precision matrix}

\item{include_x}{logical}

\item{include_y}{logical}
}
\value{
a matrix of imputed values
}
\description{
MCMC sample of missing data
}
