% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{update_z}
\alias{update_z}
\title{MCMC sample of the n latent variables in the probit model (z)}
\usage{
update_z(Y, m, B, R)
}
\arguments{
\item{Y}{matrix of probit covariates}

\item{m}{matrix of missing data indicators}

\item{B}{matrix of probit parameters}

\item{R}{correlation matrix}
}
\value{
one MCMC sample of z
}
\description{
MCMC sample of the n latent variables in the probit model (z)
}
