% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBART.R
\name{mvBART}
\alias{mvBART}
\title{Title}
\usage{
mvBART(
  x,
  y,
  x_predict = NA,
  n_trees = 100,
  burn = 1000,
  iters = 1000,
  thin = 2,
  predict = TRUE,
  tree_prior_params = tree_list(...),
  hypers = hypers_list(...),
  scale = TRUE,
  show_progress = TRUE,
  progress_every = 10,
  true_trees_data = NA,
  ...
)
}
\arguments{
\item{x}{covariates}

\item{y}{response}

\item{x_predict}{out-of-sample covariates}

\item{n_trees}{number of trees}

\item{burn}{burn-in samples}

\item{iters}{post burn-in samples}

\item{thin}{thinning}

\item{predict}{whether or not to make out-of-sample predictions}

\item{tree_prior_params}{prior parameters for BART trees}

\item{hypers}{prior parameters for BART parameters}

\item{scale}{logical. Whether to scale data to range (-0.5, 0.5).}

\item{show_progress}{logical.}

\item{progress_every}{integer value stating how often to update the progress bar.}

\item{true_trees_data}{true trees for BART component}

\item{...}{Catches unused arguments}
}
\value{
An object of class \code{"BART"}.
}
\description{
Title
}
\examples{
# data <- sim_data_friedman(n = 100, p = 2)
# bart_out <- mvBART(data$x, data$y, n_trees = 90, burn = 500,
#                    iters = 1000, thin = 2, predict = FALSE)
}
