% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{propose_tree}
\alias{propose_tree}
\title{Title}
\usage{
propose_tree(df, x, x_vars, min_node, max_attempt = 100, i)
}
\arguments{
\item{df}{dataframe}

\item{x}{covariates}

\item{x_vars}{vector}

\item{min_node}{minimum number of observations that should fall into a single terminal node.}

\item{max_attempt}{maximum number of attempts to find a suitable tree}

\item{i}{iteration number}
}
\description{
Title
}
