% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{propose_tree}
\alias{propose_tree}
\title{A function for proposing an updated tree structure}
\usage{
propose_tree(
  df,
  x,
  min_node,
  max_attempt = 10,
  i,
  probit = FALSE,
  miss_row = NA,
  cat_list = NA
)
}
\arguments{
\item{df}{dataframe}

\item{x}{covariates}

\item{min_node}{minimum number of observations that should fall into a single terminal node.}

\item{max_attempt}{maximum number of attempts to find a suitable tree}

\item{i}{iteration number}

\item{probit}{logical. Indicates whether tree is for continuous or binary outcomes.}

\item{miss_row}{index of missing rows}

\item{cat_list}{list of used categorical subset}
}
\value{
A data.frame containing details of a new tree structure
}
\description{
A function for proposing an updated tree structure
}
\examples{
# Create a data.frame representing the root node of a tree
df <- data.frame(matrix(ncol = 8, nrow = 1))
colnames(df) <- c("parent", "lower", "upper", "split_variable",
                  "split_value", "depth", "direction", "NA_direction")
df[1,] <- c(0,0,1,0,1,0,0, NA)
x <- matrix(stats::runif(9), ncol=3)
propose_tree(df, x, min_node = 1, max_attempt = 1, i = 1)
}
