% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypers.R
\name{tree_list}
\alias{tree_list}
\title{Tree prior parameters}
\usage{
tree_list(prior_alpha = 0.95, prior_beta = 2, min_node = 5, max_attempt = 100)
}
\arguments{
\item{prior_alpha}{hyperparameter}

\item{prior_beta}{hyperparameter}

\item{min_node}{minimum number of observations that should fall into a single terminal node.}

\item{max_attempt}{maximum number of attempts to find a suitable tree}
}
\value{
list of hyperparameters
}
\description{
Tree prior parameters
}
\examples{
tree_list(prior_alpha = 0.8, prior_beta = 3)
}
